refactor(agentic-ai): trim verbose Bedrock Converse javadoc, move v1/v2 stub split out of the shared base
Cuts class/method javadoc down to facts not already visible from the code
(stop-reason handling, event-frame ordering, prompt-caching rejection rule),
and aligns the two tool-result 'unsupported content type' error messages.

AbstractBedrockConverseWireFormatFixture's stubConversation() implemented
the v1 (buffered JSON) shape even though the SPI already declares the
method abstract and v2 already overrides it; moved that implementation
into BedrockConverseV1WireFormatFixture so the shared base only holds what
is actually shared between the two fixtures.
