+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/aiagent/wiremock/bedrock/AbstractBedrockConverseWireFormatFixture.java
@@ -16,12 +16,8 @@
  */
 package io.camunda.connector.e2e.agenticai.aiagent.wiremock.bedrock;
 
-import io.camunda.connector.e2e.agenticai.aiagent.wiremock.bedrock.BedrockConverseChatModelStubs.ToolCall;
-import io.camunda.connector.e2e.agenticai.aiagent.wiremock.bedrock.BedrockConverseChatModelStubs.Turn;
 import io.camunda.connector.e2e.agenticai.aiagent.wiremock.spi.ProviderWireFormatFixture;
 import io.camunda.connector.e2e.agenticai.aiagent.wiremock.spi.RecordedChatRequest;
-import io.camunda.connector.e2e.agenticai.aiagent.wiremock.spi.TurnStub;
-import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -29,8 +25,8 @@ import java.util.List;
  * ({@link BedrockConverseV1WireFormatFixture}, plain buffered JSON via {@code converse}) and
  * streaming ({@link BedrockConverseV2WireFormatFixture}, AWS EventStream via {@code
  * converseStream}) fixtures — the request wire format and the recorded-request parsing are
- * identical between the two; only the response framing differs (see {@link
- * #stubConversation(TurnStub...)}, overridden by the v2 fixture).
+ * identical between the two; only the response framing differs, so each fixture stubs its own
+ * conversation independently.
  *
  * <p>Notable wire-level differences from OpenAI/Anthropic:
  *
@@ -52,28 +48,6 @@ abstract class AbstractBedrockConverseWireFormatFixture implements ProviderWireF
   @Override
   public String toString() {
     return apiName();
-  }
-
-  @Override
-  public void stubConversation(TurnStub... turns) {
-    BedrockConverseChatModelStubs.stubConversation(
-        Arrays.stream(turns)
-            .map(AbstractBedrockConverseWireFormatFixture::toStubTurn)
-            .toArray(Turn[]::new));
-  }
-
-  private static Turn toStubTurn(TurnStub turn) {
-    return switch (turn) {
-      case TurnStub.Text text -> Turn.text(text.text(), text.inputTokens(), text.outputTokens());
-      case TurnStub.ToolCalls toolCalls ->
-          Turn.toolCalls(
-              toolCalls.text(),
-              toolCalls.inputTokens(),
-              toolCalls.outputTokens(),
-              toolCalls.toolCalls().stream()
-                  .map(tc -> ToolCall.of(tc.id(), tc.name(), tc.argumentsJson()))
-                  .toArray(ToolCall[]::new));
-    };
   }
 
   @Override
