+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/aiagent/wiremock/bedrock/BedrockConverseV1WireFormatFixture.java
@@ -18,22 +18,40 @@ package io.camunda.connector.e2e.agenticai.aiagent.wiremock.bedrock;
 
 import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo;
 import io.camunda.connector.e2e.ElementTemplate;
-import io.camunda.connector.e2e.agenticai.aiagent.wiremock.spi.ProviderWireFormatFixture;
+import io.camunda.connector.e2e.agenticai.aiagent.wiremock.bedrock.BedrockConverseChatModelStubs.ToolCall;
+import io.camunda.connector.e2e.agenticai.aiagent.wiremock.bedrock.BedrockConverseChatModelStubs.Turn;
+import io.camunda.connector.e2e.agenticai.aiagent.wiremock.spi.TurnStub;
+import java.util.Arrays;
 import java.util.function.Function;
 
-/**
- * Plugs AWS Bedrock's Converse API wire format into the provider-agnostic {@link
- * ProviderWireFormatFixture} SPI, driving the connector through the v1 element template (the
- * LangChain4j-based provider, a plain, non-streaming {@code converse} call). See {@link
- * AbstractBedrockConverseWireFormatFixture} for the wire-format plumbing shared with the v2 fixture
- * ({@link BedrockConverseV2WireFormatFixture}).
- */
 public final class BedrockConverseV1WireFormatFixture
     extends AbstractBedrockConverseWireFormatFixture {
 
   @Override
   public String apiName() {
     return "BedrockConverseV1";
+  }
+
+  @Override
+  public void stubConversation(TurnStub... turns) {
+    BedrockConverseChatModelStubs.stubConversation(
+        Arrays.stream(turns)
+            .map(BedrockConverseV1WireFormatFixture::toStubTurn)
+            .toArray(Turn[]::new));
+  }
+
+  private static Turn toStubTurn(TurnStub turn) {
+    return switch (turn) {
+      case TurnStub.Text text -> Turn.text(text.text(), text.inputTokens(), text.outputTokens());
+      case TurnStub.ToolCalls toolCalls ->
+          Turn.toolCalls(
+              toolCalls.text(),
+              toolCalls.inputTokens(),
+              toolCalls.outputTokens(),
+              toolCalls.toolCalls().stream()
+                  .map(tc -> ToolCall.of(tc.id(), tc.name(), tc.argumentsJson()))
+                  .toArray(ToolCall[]::new));
+    };
   }
 
   @Override
