+++ connectors/agentic-ai/connector-agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/chatmodel/provider/bedrock/BedrockConverseRequestConverter.java
@@ -259,15 +259,13 @@ public class BedrockConverseRequestConverter {
   }
 
   /**
-   * Finds the last message eligible for the moving checkpoint. Confirmed empirically against the
-   * real API (AWS returns "extraneous key [cachePoint] is not permitted" otherwise): Bedrock
-   * rejects a {@code cachePoint} block in any message that contains a {@code toolUse}, {@code
-   * toolResult}, or {@code reasoningContent} block -- only messages made entirely of {@code
-   * text}/{@code image}/{@code document} content qualify. This matches every placement example in
-   * AWS's own docs, where a cachePoint only ever follows {@code text}/{@code image}. In an agentic
-   * loop this means the checkpoint cannot advance past a tool round-trip -- it stays pinned to the
-   * last plain-text message -- but it still moves for multi-turn conversations without tool calls
-   * in between.
+   * Finds the last message eligible for the moving checkpoint. Bedrock rejects a {@code cachePoint}
+   * block in any message that contains a {@code toolUse}, {@code toolResult}, or {@code
+   * reasoningContent} block ("extraneous key [cachePoint] is not permitted") -- only messages made
+   * entirely of {@code text}/{@code image}/{@code document} content qualify. In an agentic loop
+   * this means the checkpoint cannot advance past a tool round-trip -- it stays pinned to the last
+   * plain-text message -- but it still moves for multi-turn conversations without tool calls in
+   * between.
    */
   private static int lastCacheableMessageIndex(List<Message> messages) {
     for (int i = messages.size() - 1; i >= 0; i--) {
@@ -285,8 +283,6 @@ public class BedrockConverseRequestConverter {
     return -1;
   }
 
-  // No ttl is ever set: AWS's default 5-minute TTL applies, and the "longer TTL must precede
-  // shorter" ordering rule cannot be violated by a request that never specifies one.
   private static CachePointBlock defaultCachePoint() {
     return CachePointBlock.builder().type(CachePointType.DEFAULT).build();
   }
